Changed nack to requeue msg

+++ prepare.sh
@@ -1,10 +1,19 @@
 #!/bin/bash
 
 cd bin
-wget "https://downloads.rclone.org/rclone-current-linux-amd64.zip"
-unzip "rclone-current-linux-amd64.zip"
-rm "rclone-current-linux-amd64.zip"
-cd */
-mv rclone ../
+
+if [ -f rclone ]; then
+  echo "Rclone binary found"
+else
+  echo "Downloading rclone"
+  wget "https://downloads.rclone.org/rclone-current-linux-amd64.zip"
+  unzip "rclone-current-linux-amd64.zip"
+  rm "rclone-current-linux-amd64.zip"
+  cd */
+  mv rclone ../
+  cd ..
+  rm -rf rclone*/
+fi
+
 cd ..
-rm -rf rclone*/
+npm i
